pyproject.toml
[project]
name = "ai-applications"
version = "0.1.0"
description = ""
authors = [
    {name = "Piotr Sacharuk",email = "piotr.sacharuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "textblob (>=0.19.0,<0.20.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pymongo (>=4.15.4,<5.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "langchain (>=1.1.0,<2.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "awswrangler (>=3.14.0,<4.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "faiss-gpu (>=1.7.2,<2.0.0)",
    "faiss-cpu (>=1.13.0,<2.0.0)",
    "pypdf (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
